% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{benthicData}
\alias{benthicData}
\title{Import Benthic Bag Tablet Files Into DataEnt.mdb}
\usage{
benthicData(haul, dsnTablet, dsnDataEnt)
}
\arguments{
\item{haul}{haul number}

\item{dsnTablet}{full file path of the location of raw tablet data}

\item{dsnDataEnt}{full file path of the location of the dataent.mdb - include full file name and extension.}
}
\value{
...
}
\description{
NOTE: Must be run on 32-bit R!
}
\examples{
# benthicData()
}
