% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_net_spread}
\alias{calc_net_spread}
\title{Volumetric Tow Calculator}
\usage{
calc_net_spread(dat)
}
\arguments{
\item{dat}{data.frame. This can be data from GIDES or race_data.hauls, including these columns: WIRE_OUT, NET_HEIGHT, NET_SPREAD}

\item{volume_method}{Character: "average" (default) or "contour". If "contour", an inverse-distance weighted interpolation will be applied to the heights across the bin and summed to find volume.
If "average", the bin heights will be averaged in 4 area quadrants (as done in old documentation) to find volume.}

\item{density_method}{Character: "estimated" (default) or "observed". For "observed", add a numeric of the observed (calculated by you on deck) density. For "estimated", create a 2 column data.frame with "taxon" and "prop" (proportion) of each fish caught in the survey: "pollock", "pcod", "Atka mackerel", "pop", "nrf", "Giant grenadier", "flatfish", "snow crab"}

\item{bin_dimentions}{list of 3 items (length of "bow" side of shape, length of "stern" side of shape, and the "height"s between them) or NULL (default).}

\item{area_of_bin}{A single numeric with the area, or the bin in cm3 or NULL (default). If null, area will be calculated from the bin_dimensions. Will be used instead of bin_dimentions if not NULL.}

\item{heights_in_bin}{List of 9 items. See example below for structure. 9 hieghts from the bin should be taken from the stern starbord, center starbord, bow starbord, stern center, center center, bow center, stern port, center port, bow port}

\item{density}{Dataframe or numeric. For density_method = "observed", add a numeric of the observed (calculated by you on deck) density. For density_method = "estimated", create a 2 column data.frame with "taxon" and "prop" (proportion) of each fish caught in the survey: "pollock", "pcod", "Atka mackerel", "pop", "nrf", "Giant grenadier", "flatfish", "snow crab"}
}
\value{
a list of equations
}
\description{
recall on a boat...
    bow
port   starbord
   stern
thus heights in the bin should be taken like this:
                    ...bow...
bow-port (bp)       bow-center (bc)      bow-starbord (bs)
center-port (cp)    center-center (cc)   center-starbord (cs)
stern-port (sp)     stern-center (sc)    stern-starbord (ss)
                    ...stern...
}
\examples{
# Here is an example using 202101 Alaska Night Data from race_data.hauls:
path <- system.file("exdata/calc_net_spread/VESTY202101_RACE_DATA.HAULS.csv",
                    package = "GAPsurvey")
dat <- read.csv(file=path, header=TRUE, sep=",", stringsAsFactors = FALSE)
dat <- dat[dat$PERFORMANCE >= 0 & dat$HAUL_TYPE == 3,]
dat[dat$NET_SPREAD_METHOD != 7, "NET_SPREAD"] <- NA # a normal net width
dat[dat$NET_HEIGHT_METHOD != 6, "NET_HEIGHT"] <- NA # a normal net height
calc_net_spread(dat)
}
