% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_volumetric_tow}
\alias{calc_volumetric_tow}
\title{Volumetric Tow Calculator}
\usage{
calc_volumetric_tow(
  volume_method = "contour",
  density_method = "estimated",
  bin_dimentions = NULL,
  area_of_bin = NULL,
  heights_in_bin,
  density = NULL
)
}
\arguments{
\item{volume_method}{Character: "average" (default) or "contour". If "contour", an inverse-distance weighted interpolation will be applied to the heights across the bin and summed to find volume.
If "average", the bin heights will be averaged in 4 area quadrants (as done in old documentation) to find volume.}

\item{density_method}{Character: "estimated" (default) or "observed". For "observed", add a numeric of the observed (calculated by you on deck) density. For "estimated", create a 2 column data.frame with "taxon" and "prop" (proportion) of each fish caught in the survey: "pollock", "pcod", "Atka mackerel", "pop", "nrf", "Giant grenadier", "flatfish", "snow crab"}

\item{bin_dimentions}{list of 3 items (length of "bow" side of shape, length of "stern" side of shape, and the "height"s between them) or NULL (default).}

\item{area_of_bin}{A single numeric with the area, or the bin in cm3 or NULL (default). If null, area will be calculated from the bin_dimensions. Will be used instead of bin_dimentions if not NULL.}

\item{heights_in_bin}{List of 9 items. See example below for structure. 9 hieghts from the bin should be taken from the stern starbord, center starbord, bow starbord, stern center, center center, bow center, stern port, center port, bow port}

\item{density}{Dataframe or numeric. For density_method = "observed", add a numeric of the observed (calculated by you on deck) density. For density_method = "estimated", create a 2 column data.frame with "taxon" and "prop" (proportion) of each fish caught in the survey: "pollock", "pcod", "Atka mackerel", "pop", "nrf", "Giant grenadier", "flatfish", "snow crab"}
}
\description{
recall on a boat...
    bow
port   starbord
   stern
thus heights in the bin should be taken like this:
                    ...bow...
bow-port (bp)       bow-center (bc)      bow-starbord (bs)
center-port (cp)    center-center (cc)   center-starbord (cs)
stern-port (sp)     stern-center (sc)    stern-starbord (ss)
                    ...stern...
}
\examples{
# Example from EBS 2022 Alaska Knight Haul 144 where a large bag (4650 kg)
# was weighed by a load cell but volumetric measurements were also taken for
# comparison and study.

# 9 heights from the bin at each location described below
# see notes above for boat layout.
heights_in_bin <- list( # in cm
 "ss" = 65, # stern starbord
 "cs" = 55, # center starbord
 "bs" = 43, # bow starbord
 "sc" = 51, # stern center
 "cc" = 34, # center center
 "bc" = 27, # bow center
 "sp" = 20, # stern port
 "cp" = 29, # center port
 "bp" = 31) # bow port

# develop a dataframe of the taxon groups below and proportions found from
# sample bag (be sure not to include weights of organisms rescued from the bin!)

density_prop <- data.frame(taxon = c("pollock", "pcod", "Atka mackerel",
                                     "pop", "nrf", "Giant grenadier",
                                     "flatfish", "snow crab"),
                           prop = c(0.045011, 0.00533, 0,
                                    0.8843, 0, 0,
                                    0.022022, 0.0000398))
# Prefered/easiest

calc_volumetric_tow(volume_method = "contour",
                             # Using function to develop an IDW
                             # contour over the heights for estimating volume.
               density_method = "estimated",
                             # Determining density proportions to apply to
                             # known densities of fish
               bin_dimentions = # Dimensions for the Alaska Knight's
                                # trapezoidal bin, measured in cm
                                list("bow" = 420,
                                     "stern" = 355,
                                     "height" = 315.5),
                             # the height is actually 320, but subtract 4.5
                             # for width of board in middle
               heights_in_bin = heights_in_bin, # measured in cm
               density = density_prop)

# Using other methods
calc_volumetric_tow(volume_method = "average",
                       # Averaging the 4 quadrants of the bin to find volume
               density_method = "observed",
                       # Directly measured density. See documentation in RACE
                       # Survey app for more details.
               bin_dimentions = 162,
                       # Alaska Knight Vessel code - bin_dimensions stored in
                       # the function and this vessel code will look them up
               heights_in_bin = heights_in_bin, # Measured in cm
               density = 775.8343)
                       # if you measured the density directly (this happens
                       # to be what the proportions above would solve for)

# Using other methods
calc_volumetric_tow(volume_method = "average",
                       # Averaging the 4 quadrants of the bin to find volume
               density_method = "observed",
                       # Directly measured density. See documentation in
                       # RACE Survey app for more details.
               area_of_bin = 122256.2,  # calculated area of the bin.
               heights_in_bin = heights_in_bin, # Measured in cm
               density = 775.8343)
                       # if you measured the density directly (this happens
                       # to be what the proportions above would solve for)

# F/V Vesteraalen Haul 128 EBS 2022
calc_volumetric_tow( # volume_method = "average",
  bin_dimentions = 94,
  heights_in_bin = list( # in cm
    "ss" = 50, # stern starbord
    "cs" = 63, # center starbord
    "bs" = 56, # bow starbord
    "sc" = 50, # stern center
    "cc" = 60, # center center
    "bc" = 61, # bow center
    "sp" = 50, # stern port
    "cp" = 50, # center port
    "bp" = 41), # bow port,
 density = data.frame("taxon" = c("pollock", "pcod", "Atka mackerel",
    "pop", "nrf", "Giant grenadier", "flatfish", "snow crab"),
    "prop" = c(1, 0, 0, 0, 0, 0, 0, 0)))
}
