% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{catchData}
\alias{catchData}
\title{Import Catch Tablet Files Into DataEnt.mdb}
\usage{
catchData(haul, dsnTablet, dsnDataEnt, importLength = TRUE)
}
\arguments{
\item{haul}{haul number}

\item{dsnTablet}{full file path of the location of raw tablet data}

\item{dsnDataEnt}{full file path and filename of the location of the dataent.mdb - include full file name and extension.}

\item{importLength}{logical - import length data (GAPsurvey::lengthData()) prior to importing catch data}
}
\value{
...
}
\description{
NOTE: Must be run on 32-bit R!
}
\details{
Processes catch data exported from catch tablets into summarized values into the designated dataent.mdb. You must have exported from the tablet:
DATAENT_CATCH... and RAW_CATCH_HAUL... for the designated haul.
}
\examples{
# catchData()
}
