% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkFiles}
\alias{checkFiles}
\title{Check files before import}
\usage{
checkFiles(haul, dsnTablet, dsnDataEnt)
}
\arguments{
\item{haul}{Number of the haul you want to upload}

\item{dsnTablet}{String for the drive folder (in R notation) where tablet files exist.}

\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}
}
\value{
Print statements indicating if any raw data files are missing and which length tablet data are being used.
}
\description{
Check files before import
}
\examples{
dsnTablet <- system.file("exdata/catch/GOA/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/GOA/data_ent.mdb", package = "GAPsurvey")

# Define the current haul number
haul <- 8

# Import length data
checkFiles(haul, dsnTablet, dsnDataEnt)
}
