% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_ctd_hex}
\alias{convert_ctd_hex}
\title{Convert CTD .hex file to .cnv using SBE Data Processing}
\usage{
convert_ctd_hex(
  hex_file_path,
  xmlcon_path,
  bat_file = NULL,
  datcnv_file = NULL
)
}
\arguments{
\item{hex_file_path}{Path to a single .hex file from the CTD.}

\item{xmlcon_path}{Path to CTD the config file (.xmlcon) for the CTD that was used to collect data in the .hex file.}

\item{bat_file}{Optional. Filepath to a batch (.bat) file to be executed in command line to run SBE Data Processing. Default NULL automatically uses the .bat file that is included in GAPsurvey.}

\item{datcnv_file}{Optional. Filepath to .psa file for the SBE Data Processing Datcnv module. Default NULL automatically uses the .psa file that is included in GAPsurvey.}
}
\value{
Writes decoded CTD data to a cnv file and returns the path to the file as a 1L character vector.
}
\description{
Converts SBE19plus V2 CTD hex file to cnv using Sea-Bird Data Processing software. cnv files generated with this function will be properly formatted for conversion to BTD using convert_ctd_btd().
}
\examples{
# Copy system files to working directory for example
file.copy(from = system.file("exdata/convert_ctd_btd/2021_06_13_0003.hex",
package = "GAPsurvey"),
to = gsub(pattern = system.file("exdata/convert_ctd_btd/", package = "GAPsurvey"),
          replacement = getwd(),
          x = system.file("exdata/convert_ctd_btd/2021_06_13_0003.hex",
          package = "GAPsurvey")))

file.copy(from = system.file("exdata/convert_ctd_btd/19-8102_Deploy2021.xmlcon",
                             package = "GAPsurvey"),
          to = gsub(pattern = system.file("exdata/convert_ctd_btd/",
                                           package = "GAPsurvey"),
                                           replacement = getwd(),
                    x = system.file("exdata/convert_ctd_btd/19-8102_Deploy2021.xmlcon",
                                    package = "GAPsurvey")))

# Run convert_ctd_hex()
GAPsurvey::convert_ctd_hex(hex_file_path = "2021_06_13_0003.hex",
                            xmlcon_path = "19-8102_Deploy2021.xmlcon",
                            bat_file = NULL,
                            datcnv_file = NULL)
}
