% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CTDtoBTD}
\alias{CTDtoBTD}
\title{Convert raw CTD files to BTD/BTH format}
\usage{
CTDtoBTD(
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  DATE = NA,
  MODEL_NUMBER = NA,
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA,
  path_in = NA,
  path_out = "./",
  filename_add = "",
  quiet = FALSE
)
}
\arguments{
\item{VESSEL}{Optional. Default = NA. The vessel number (e.g., 94). If NA or not called in the function, a prompt will appear asking for this data.}

\item{CRUISE}{Optional. Default = NA. The cruise number, which is usually the year date (e.g., 201901). If NA or not called in the function, a prompt will appear asking for this data.}

\item{HAUL}{Optional. Default = NA. The haul number, aka the iterative number of this haul (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.}

\item{DATE}{Optional. Default = NA. The date in MM/DD/YYYY format (e.g., "06/02/2019"). If NA or not called in the function, a prompt will appear asking for this data.}

\item{MODEL_NUMBER}{Optional. Default = NA. The model number of the CTD (I think) (e.g., 123) If NA or not called in the function, a prompt will appear asking for this data. # TOLEDO}

\item{VERSION_NUMBER}{Optional. Default = NA. The version number of the CTD (I think) (e.g., 123) If NA or not called in the function, a prompt will appear asking for this data. # TOLEDO}

\item{SERIAL_NUMBER}{Optional. Default = NA. The serial number of the CTD (I think) (e.g., 123) If NA or not called in the function, a prompt will appear asking for this data. # TOLEDO}

\item{path_in}{Optional. Default = "./., or the local working directory but any path (as a string) may be entered.}

\item{path_out}{Optional. The default is the local working directory but may be specified with a string.}

\item{filename_add}{Optional. Default = "new". This string will be added to the name of the outputed file. Here, you can additional information that may make this file helpful to find later.}

\item{quiet}{Optional logical TRUE/FALSE. Default = FALSE. If FALSE, will print a statement or a pop-up window will let the user know where the file has been saved to.}
}
\value{
A .BTH and .BTD file to the path_out directory.


}
\description{
To RUN a function, put curser on line you want to run and press CTRL+R
}
\details{
When first open "Survey R Functions" the function library("tcltk")
MUST be run ONCE
As long as R console remains open, the function library("tcltk")
does not need to be run a second time

library("tcltk")

The "CTDtoBTD()" function converts an exported CSV file from the Teledyne CTD Program into BTD and BTH files
File format for the CSV file must be DATE,TIME,TEMP,DEPTH (mm-dd-yyyy,hh:mm:ss.ss,(C),(m))
Program prompts you for: Vessel code, Cruise number, Haul number, CTD model number,
Version number (=999), CTD serial number, and then opens window for you  to select file from any directory.
The BTH and BTD files are written to the C: \ NM directory
}
\examples{
path_in0 <- system.file("exdata", package = "GAPsurvey")
# CTDtoBTD(
#     VESSEL = 94,
#     CRUISE = 201901,
#     HAUL = 3,
#     DATE = "06/02/2019",
#     MODEL_NUMBER = 123,
#     VERSION_NUMBER = 123,
#     SERIAL_NUMBER = 123,
#     path_in = paste0(path_in0, "ctd2btd/6-02-19/DATA01.ctd"),
#     path_out = path_out0,
#     filename_add = "", # is this important?
#     quiet = TRUE)
}
