% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deleteDataEnt}
\alias{deleteDataEnt}
\title{Delete data from DataEnt.mdb}
\usage{
deleteDataEnt(dsnDataEnt, haul, tablename)
}
\arguments{
\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}

\item{haul}{Number of the haul you want to delete}

\item{tablename}{String of the data_ent table you are deleting data from}
}
\value{
...
}
\description{
NOTE: Must be run on 32-bit R!
}
\details{
Deletes data from data_ent.mdb for a haul.
}
\examples{
# deleteDataEnt(dsnDataEnt,haul,"BENTHIC_BAG")
}
