% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fix_path}
\alias{fix_path}
\title{Make sure file path is complete}
\usage{
fix_path(path)
}
\arguments{
\item{path}{A string with the complete path of the directory or file.}
}
\value{
A fixed path string.
}
\description{
Function adds '/' or '\\' to the end of directories and recognizes when there are file extentions at the end of strings.
}
\examples{
fix_path("sdfg/sdfg/sdfg/dfg.dd")
fix_path("sdfg/sdfg/sdfg")
fix_path("sdfg/sdfg/sdfg/")
}
