% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lengthData}
\alias{lengthData}
\title{Import Length Tablet Files Into DataEnt.mdb}
\usage{
lengthData(haul, dsnTablet, dsnDataEnt)
}
\arguments{
\item{haul}{Number of the haul you want to upload}

\item{dsnTablet}{String for the drive folder (in R notation) where tablet files exist.}

\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}
}
\value{
changes to catch data in the access database (.mbd)
}
\description{
NOTE: Must be run on 32-bit R!
NOTE: This function is typically run within the catchData()
}
\details{
Looks for tablet length files in given DSN and uploads them to LENGTH and RAW_LENGTH in the given data_ent.mdb.
}
\examples{
# Gulf of Alaska Example
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/GOA/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/GOA/data_ent.mdb", package = "GAPsurvey")

# Define the current haul number
haul <- 8

# Import length data
lengthData(haul,dsnTablet,dsnDataEnt)  



# Eastern Bering Sea Example
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/EBS/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/EBS/data_ent.mdb", package = "GAPsurvey")

# Define the current haul number
haul <- 59

# Import length data
lengthData(haul,dsnTablet,dsnDataEnt) 
}
