% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{local_racebase}
\alias{local_racebase}
\title{RACEBASE Data Set}
\format{
An object of class \code{data.frame} with 1613690 rows and 38 columns.
}
\usage{
local_racebase
}
\description{
RACEBASE Data Set
}
\examples{
data(local_racebase)
}
\keyword{datasets}
