% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_install.R
\name{pkg_install}
\alias{pkg_install}
\title{Package install}
\usage{
pkg_install(PKG)
}
\arguments{
\item{PKG}{name of package
this script has additional capability to work on google workstations. 
You can use the library install bash script.}
}
\description{
Package install
}
\keyword{internal}
