% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{PolySpecies}
\alias{PolySpecies}
\title{PolySpecies Data Set}
\format{
An object of class \code{data.frame} with 172 rows and 4 columns.
}
\usage{
PolySpecies
}
\description{
PolySpecies Data Set
}
\examples{
data(PolySpecies)
}
\keyword{datasets}
