% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{selectDataEnt}
\alias{selectDataEnt}
\title{Query DataEnt.mdb}
\usage{
selectDataEnt(dsnDataEnt, query)
}
\arguments{
\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}

\item{query}{An SQL quiery # TOLEDO}
}
\value{
...
}
\description{
NOTE: Must be run on 32-bit R!
}
\details{
Executes SQL queries of data_ent.mdb
}
\examples{
# selectDataEnt()
}
