% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{sp_table}
\alias{sp_table}
\title{species data codes}
\format{
A data frame with 2754 rows and 2 variables:
\describe{
\item{\code{species_code}}{integer Species code}
\item{\code{report_name_scientific}}{Species scientific names}
}
}
\usage{
data(sp_table)
}
\description{
RACEBASE Species Codes and Scientific Names Data Set
}
\details{
DETAILS
}
\examples{
utils::data(sp_table)
}
\author{
Margaret Siple (margaret.siple AT noaa.gov)
}
\keyword{code}
\keyword{data}
\keyword{scientific}
\keyword{species}
