% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{specimenData}
\alias{specimenData}
\title{Import Specimen Tablet Files Into DataEnt.mdb}
\usage{
specimenData(haul, dsnTablet, dsnDataEnt)
}
\arguments{
\item{haul}{Number of the haul you want to upload}

\item{dsnTablet}{String for the drive folder (in R notation) where tablet files exist.}

\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}
}
\value{
...
}
\description{
Looks for tablet specimen files in given DSN and uploads them to SPECIMEN in the given data_ent.mdb.
}
\examples{
# specimenData()
}
