% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TEDtoBTD}
\alias{TEDtoBTD}
\title{BVDR Conversion to Create BTD data}
\usage{
TEDtoBTD(
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  MODEL_NUMBER = NA,
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA,
  path_in = "C:/Program Files/Marport Server/Logs/",
  path_out = "./",
  filename_add = "new",
  quiet = FALSE
)
}
\arguments{
\item{VESSEL}{Optional. Default = NA. The vessel number (e.g., 94). If NA or not called in the function, a prompt will appear asking for this data.}

\item{CRUISE}{Optional. Default = NA. The cruise number, which is usually the year date (e.g., 201901). If NA or not called in the function, a prompt will appear asking for this data.}

\item{HAUL}{Optional. Default = NA. The haul number, aka the iterative number of this haul (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.}

\item{MODEL_NUMBER}{Optional. Default = NA. The model number of the Marport reciever (I think) (e.g., 123) If NA or not called in the function, a prompt will appear asking for this data. # TOLEDO}

\item{VERSION_NUMBER}{Optional. Default = NA. The version number of the Marport reciever (I think) (e.g., 123) If NA or not called in the function, a prompt will appear asking for this data. # TOLEDO}

\item{SERIAL_NUMBER}{Optional. Default = NA. The serial number of the Marport reciever (I think) (e.g., 123) If NA or not called in the function, a prompt will appear asking for this data. # TOLEDO}

\item{path_in}{Optional. The default is the localion on the catch computer ("C:/Program Files/Marport Server/Logs/") but any path can be entered.}

\item{path_out}{Optional. The default is the local working directory but can be specified with a string.}

\item{filename_add}{Optional. Default = "new". This string will be added to the name of the outputed file. Here, you can additional information that may make this file helpful to find later.}

\item{quiet}{Optional logical TRUE/FALSE. Default = FALSE. If FALSE, will print a statement or a pop-up window will let the user know where the file has been saved to.}
}
\value{
A .BTH and .BTD file to the path_out directory.
}
\description{
Converts Marport BVDR data (TED) to RACE BTD format.
}
\examples{
path_in0 <- system.file("exdata", package = "GAPsurvey")
TEDtoBTD(
   VESSEL = 94,
   CRUISE = 201901,
   HAUL = 3,
   MODEL_NUMBER = 123,
   VERSION_NUMBER = 123,
   SERIAL_NUMBER = 123,
   path_in = paste0(path_in0, "/bvdr2btd/"),
   path_out = getwd(),
   filename_add = "new", # is this important?
   quiet = TRUE)
}
