% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_to_cnv.R
\name{write_to_cnv}
\alias{write_to_cnv}
\title{Writes outputs of hex_to_cnv to a .cnv file}
\usage{
write_to_cnv(data_list, output_path)
}
\arguments{
\item{data_list}{List of data and metadata to write to the cnv file}

\item{output_path}{Output path for the cnv file}
}
\description{
Appends header data.
}
\author{
Sean Rohan
}
