% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{writeDataEnt}
\alias{writeDataEnt}
\title{Write records to DataEnt.mdb}
\usage{
writeDataEnt(dsnDataEnt, data, tablename)
}
\arguments{
\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}

\item{data}{matrix or data.frame containg data from catch, specimen and lengths tablets}

\item{tablename}{String of the data_ent table you are deleting data from}
}
\value{
...
}
\description{
NOTE: Must be run on 32-bit R!
}
\details{
Creates connection channel to data_ent.mdb, then appends data to given table.
}
\examples{
# writeDataEnt()
}
