% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{benthicData}
\alias{benthicData}
\title{Import Benthic Bag Tablet Files Into DataEnt.mdb}
\usage{
benthicData(haul, dsnTablet, dsnDataEnt)
}
\arguments{
\item{haul}{haul number}

\item{dsnTablet}{full file path of the location of raw tablet data}

\item{dsnDataEnt}{full file path of the location of the dataent.mdb - include full file name and extension.}
}
\value{
changes to catch data in the access database (.mbd)
}
\description{
NOTE: Must be run on 32-bit R!
}
\examples{
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/GOA/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/GOA/data_ent.mdb", package = "GAPsurvey")
importLength <- TRUE ## Change this "T" to "F" if not importing length data ##

# Define the current haul number
haul <- 8

# Import catch data
# benthicData(haul,dsnTablet,dsnDataEnt)
}
