% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_net_spread.R, R/functions.R
\name{calc_net_spread}
\alias{calc_net_spread}
\title{Calculate Net Spread for tows missing net width using a glm.}
\usage{
calc_net_spread(dat)

calc_net_spread(dat)
}
\arguments{
\item{dat}{data.frame. This can be data from GIDES or race_data.hauls, including these columns: WIRE_OUT, NET_HEIGHT, NET_SPREAD}
}
\value{
a list of equations

a list of equations
}
\description{
The Marport Deep Sea Technologies Inc. net mensuration system was used during the deployment of each tow to record net spread and net height. Net width was measured as the horizontal distance between two sensors attached immediately forward of the junction of the upper breastline and the dandyline, and net height was measured from the headrope center to the seafloor. A custom-made AFSC bottom contact sensor (accelerometer) attached to the center of the footrope was used to determine tow duration based on footrope contact with the seafloor. Mean calc_net_spread values for estimating area swept for the tow duration were calculated according to the methods described by Lauth and Kotwicki (2014).
}
\details{
In race_data, this will manifest as...
net_mensuration_code = Net Mensuration Method
0* Unidentified method. Will make racebase.haul$net_mesured = "N".
1 Scanmar net mensuration - don't use, historical
2 NetMind net mensuration - don't use, historical
3 Furuno net mensuration - don't use, historical
4* Estimated from other hauls - when missing net spread, or spread and hieght. Will make racebase.haul$net_mesured = "N". Estimated using GLM
5* Estimated from warp angle - hopefully, will not come up and shouldn't be used
6 Marport net mensuration - shouldn't be there, indicates raw data
7* Marport with sequential outlier rejection, smoothed mean, and adjusted for MKII offset (see AFSC Proc. Report Lauth & Kotwicki 2014). Will make racebase.haul$net_mesured = "Y".

The Marport Deep Sea Technologies Inc. net mensuration system was used during the deployment of each tow to record net spread and net height. Net width was measured as the horizontal distance between two sensors attached immediately forward of the junction of the upper breastline and the dandyline, and net height was measured from the headrope center to the seafloor. A custom-made AFSC bottom contact sensor (accelerometer) attached to the center of the footrope was used to determine tow duration based on footrope contact with the seafloor. Mean calc_net_spread values for estimating area swept for the tow duration were calculated according to the methods described by Lauth and Kotwicki (2014).
In race_data, this will manifest as...
\describe{
 \item{"net_mensuration_code"}{Net Mensuration Method}
 \item{"0*"}{Unidentified method. Will make racebase.haul$net_measured = "N"}
 \item{"1"}{Scanmar net mensuration - don't use, historical}
 \item{"2"}{NetMind net mensuration - don't use, historical}
 \item{"3"}{Furuno net mensuration - don't use, historical}
 \item{"4*"}{Estimated from other hauls - when missing net spread, or spread and height. Will make racebase.haul$net_measured = "N". Estimated using GLM}
 \item{"5*"}{Estimated from warp angle - hopefully, will not come up and shouldn't be used}
 \item{"6"}{Marport net mensuration - shouldn't be there, indicates raw data}
 \item{"7*"}{Marport with sequential outlier rejection, smoothed mean, and adjusted for MKII offset (see AFSC Proc. Report Lauth & Kotwicki 2014). Will make racebase.haul$net_measured = "Y".}
}
}
\examples{
# Here is an example using 202101 Alaska Night Data from race_data.hauls:
path <- system.file("exdata/calc_net_spread/VESTY202101_RACE_DATA.HAULS.csv", package = "GAPsurvey")
dat <- read.csv(file=path, header=TRUE, sep=",", stringsAsFactors = FALSE)
dat <- dat[dat$PERFORMANCE >= 0 & dat$HAUL_TYPE == 3,]
dat[dat$NET_SPREAD_METHOD != 7, "NET_SPREAD"] <- NA # a normal net width
dat[dat$NET_HEIGHT_METHOD != 6, "NET_HEIGHT"] <- NA # a normal net height
calc_net_spread(dat)
# Here is an example using 202101 Alaska Night Data from race_data.hauls:
path <- system.file("exdata/calc_net_spread/VESTY202101_RACE_DATA.HAULS.csv",
                    package = "GAPsurvey")
dat <- read.csv(file=path, header=TRUE, sep=",", stringsAsFactors = FALSE)
dat <- dat[dat$PERFORMANCE >= 0 & dat$HAUL_TYPE == 3,]
dat[dat$NET_SPREAD_METHOD != 7, "NET_SPREAD"] <- NA # a normal net width
dat[dat$NET_HEIGHT_METHOD != 6, "NET_HEIGHT"] <- NA # a normal net height
calc_net_spread(dat)
}
