% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{catchData}
\alias{catchData}
\title{Import Catch Tablet Files Into DataEnt.mdb}
\usage{
catchData(haul, dsnTablet, dsnDataEnt, importLength = TRUE)
}
\arguments{
\item{haul}{haul number}

\item{dsnTablet}{full file path of the location of raw tablet data}

\item{dsnDataEnt}{full file path and filename of the location of the dataent.mdb - include full file name and extension.}

\item{importLength}{logical - import length data (GAPsurvey::lengthData()) prior to importing catch data}
}
\value{
changes to catch data in the access database (.mbd)
}
\description{
Processes catch data exported from catch tablets into summarized values into the designated dataent.mdb. You must have exported from the tablet:
DATAENT_CATCH... and RAW_CATCH_HAUL... for the designated haul.
}
\details{
NOTE: Must be run on 32-bit R!

To use the program you must first:
Bluetooth transfer the following files from tablets to catch computer (should automatically go into C:Users/NOAADATA/Documents/Bluetooth/inbox)
Each of these file names will begin with the tablet name and end with the haul number.
DATAENT_CATCH_xxxx.csv
DATAENT_BBAG_xxxx.csv     -- If there was a benthic bag
RAW_BBAG_xxxx.csv         -- If there was a benthic bag
RAW_CATCH_HAUL_xxxx.csv
RAW_CATCH_HAUL_ATTR_xxxx.csv
RAW_CATCH_SAMPLE_xxxx.csv
RAW_CATCH_VALUE_xxxx.csv
SPECIMEN_xxxx.csv         -- The specimen data
HAUL_xxxx.csv             -- A length file for each tablet

PLEASE NOTE: You cannot have any species in your catch data files that have a blank subsample_weight!  If you erroneously tap on a species during data collection, you must delete that species from the tablet before transfering the data to the catch computer!

IF the catch was subsampled, the user will be given an option to define the subsample type (1 = 100\% Processed (No subsample), 2 = Subsampled, Load cell estimate of total weight, 4 = Subsampled, Visual estimate of total weight, 6 = Subsampled, basket weights of both fractions (table split), 9 = Non-quantitative catch sampling, 12 = Volumetric method. Density from Density Lookup Table, 13 = Volumetric method. Density calculated on deck from haul sample).
}
\examples{
# Gulf of Alaska Example (without error)
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/GOA/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/GOA/data_ent.mdb", package = "GAPsurvey")
importLength <- TRUE 

# Define the current haul number
haul <- 8

# Import catch data 
catchData(haul,dsnTablet,dsnDataEnt,importLength)

# Delete data for next example run
dMix <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                       tablename = 'MIXTURE')
dMixHead <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                         tablename = 'MIXTURE_HEADER')
dCatch <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                         tablename = 'CATCH')
dCatchHead <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                           tablename = 'CATCH_HEADER')
dLength <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                        tablename = 'LENGTH')
dRawLength <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                            tablename = 'RAW_LENGTH')

# Eastern Bering Sea Example (Will cause error)
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/EBS/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/EBS/data_ent.mdb", package = "GAPsurvey")
importLength <- TRUE 

# Define the current haul number
haul <- 59

# Import catch data (This will error)
# catchData(haul,dsnTablet,dsnDataEnt,importLength)
# Row 33 for "empty bivalve shells" requires a value for SUBSAMPLE_WEIGHT. 
# Return to the tablet and edit the data there before running it through 
# this function again. 
}
