% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_ctd_btd}
\alias{convert_ctd_btd}
\title{Convert CTD data in .cnv form to BTD and BTH}
\usage{
convert_ctd_btd(
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  MODEL_NUMBER = NA,
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA,
  path_in,
  path_out = "./",
  path_xmlcon = NULL,
  filename_add = ""
)
}
\arguments{
\item{VESSEL}{Required. Default = NA. The vessel number (e.g., 94). If NA or not called in the function, a prompt will appear asking for this data.}

\item{CRUISE}{Required. Default = NA. The cruise number, which is usually the year date (e.g., 201901). If NA or not called in the function, a prompt will appear asking for this data.}

\item{HAUL}{Required. Default = NA. The haul number, aka the iterative number of this haul (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.}

\item{MODEL_NUMBER}{Optional. Default = NA. The model number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).}

\item{VERSION_NUMBER}{Optional. Default = NA. The version number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).}

\item{SERIAL_NUMBER}{Optional. Default = NA. The serial number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).}

\item{path_in}{Required. Filepath to the SBE19plus hexadecimal (.hex) file from a single cast as a character vector (e.g. "C:/CTD/202301_162_L1/sbe19plus01908091_05_04_0001.hex")/.}

\item{path_out}{Optional. The default is the local working directory but may be specified with a string.}

\item{path_xmlcon}{Required. Filepath to the CTD configuration file (.xmlcon) for the CTD used for the cast (e.g. serial number 8091 would use the file with 8091 in the name ("C:/CTD/xmlcon configuration files/SBE19plusV2_8091.xmlcon").}

\item{filename_add}{Optional. Default = "new". This string will be added to the name of the outputed file. Here, you can additional information that may make this file helpful to find later.}
}
\value{
.BTH and .BTD files to the path_out directory.
}
\description{
This function uses SBE Data Processing software to convert a CTD binary-formatted .hex (hexadecimal) file to bathythermic data (.btd) and bathythermic header files (.bth). If you are unable to convert your file, please contact sean.rohan@noaa.gov.
}
\details{
If SBE Data Processing software is not installed on the same computer as the GAPsurvey package or if you encounter errors, you may need to convert the .hex file manual using SBE Data Processing software:
1. use the SBE Data Processing Program, and in the "Run" menu select "1. Data conversion".
2. Under "Program setup file" select the .psa file (should be DataCnv.psa located in CTD folder), under "Instrument configuration file" select the .xmlcon file for the CTD located in the
deployment_xmlcon folder- this is CTD specific, know your CTD's serial number, and under "Input directory" select the .hex file (from the CTD computer) that is specific to the haul that you are missing data.  Click "Start Process".
The conversion program will ask you for outputs. At the prompt, Select and Add "Time, Elapsed, seconds", "Depth, salt water, m", "Temperature, ITS-90, deg C", "Pressure, Strain Gauge, db", and "Conductivity, S/m".
3. This .cnv file must include columns for "Time, Elapsed, seconds", "Depth , salt water, m", "Temperature, ITS-90, deg C", "Pressure, Strain Gauge, db", and "Conductivity, S/m".
4. Press "Start Process" button again and a .cnv file should appear in your selected output directory.  The .cnv file can then be used for this convert_ctd_btd() function!
5. Look in your output directory (usually the "Documents" folder, but you can find it using getwd()) for your new .BTD and .BTH files.
}
\examples{
# Convert directly .hex to .btd and .bth using convert_ctd_btd to run SBE Data Processing.
GAPsurvey::convert_ctd_btd(
    path_in = system.file("exdata/convert_ctd_btd/2021_06_13_0003.hex",
                          package = "GAPsurvey"),
    path_xmlcon = system.file("exdata/convert_ctd_btd/19-8102_Deploy2021.xmlcon",
                              package = "GAPsurvey"))

convert_ctd_btd(
   VESSEL = 94,
   CRUISE = 202101,
   HAUL = 107,
   MODEL_NUMBER = "", # for example, also can make up
   VERSION_NUMBER = "", # for example, also can make up
   SERIAL_NUMBER = 8105, # for example, also can make up
   path_in = system.file(paste0("exdata/convert_ctd_btd/",
     "2021_06_13_0003.hex"),
      package = "GAPsurvey"),
   path_xmlcon = system.file(paste0("exdata/convert_ctd_btd/",
     "19-8102_Deploy2021.xmlcon"),
      package = "GAPsurvey"),
   path_out = getwd(),
   filename_add = "newctd")

# Convert a .cnv file (after .hex files has already converted using SBE Data Processing)
convert_ctd_btd(
   VESSEL = 94,
   CRUISE = 2001,
   HAUL = 4,
   MODEL_NUMBER = 123, # for example, also can make up
   VERSION_NUMBER = 456, # for example, also can make up
   SERIAL_NUMBER = 789, # for example, also can make up
   path_in = system.file(paste0("exdata/convert_ctd_btd/",
      "SBE19plus_01908103_2021_06_01_94_0004_raw.cnv"),
       package = "GAPsurvey"),
   path_xmlcon = system.file(paste0("exdata/convert_ctd_btd/",
     "19-8102_Deploy2021.xmlcon"),
      package = "GAPsurvey"),
   path_out = getwd(),
   filename_add = "newctd")
}
