% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_nmea_btd}
\alias{convert_nmea_btd}
\title{Extract depth and temperature from NMEA strings}
\usage{
convert_nmea_btd(
  nmea_strings,
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  MODEL_NUMBER = "Marport Trawl Explorer",
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA
)
}
\arguments{
\item{nmea_strings}{Character vector of NMEA strings (e.g., in a .marp file.)}

\item{VESSEL}{Optional. Default = NA. The vessel number (e.g., 162 for AK Knight, 94 for Vesteraalen). If NA or not called in the function, a prompt will appear asking for this data.}

\item{CRUISE}{Optional. Default = NA. The cruise number, which is usually the year + sequential two digit cruise (e.g., 202101). If NA or not called in the function, a prompt will appear asking for this data.}

\item{HAUL}{Optional. Default = NA. The haul number that you are trying to convert data for (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.}

\item{MODEL_NUMBER}{Optional. Default = "Marport Trawl Explorer". The model name/number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).}

\item{VERSION_NUMBER}{Optional. Default = NA. The version number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual version number without any negative repercussions).}

\item{SERIAL_NUMBER}{Optional. Default = NA. The serial number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual serial number without any negative repercussions).}

\item{INSTRUMENT}{Instrument name as a character vector. Default = "Marport TE"}
}
\description{
Convert Marport Trawl Explorer NMEA strings to BTD and BTH files. Called internally by convert_bvdr_marp().
}
\author{
Sean Rohan <sean.rohan@noaa.gov>
}
