% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CTDtoBTD}
\alias{CTDtoBTD}
\title{Convert CTD data in .cnv form to BTD and BTH}
\usage{
CTDtoBTD(
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  MODEL_NUMBER = NA,
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA,
  path_in = NA,
  path_out = "./",
  filename_add = "",
  quiet = FALSE
)
}
\arguments{
\item{VESSEL}{Optional. Default = NA. The vessel number (e.g., 94). If NA or not called in the function, a prompt will appear asking for this data.}

\item{CRUISE}{Optional. Default = NA. The cruise number, which is usually the year date (e.g., 201901). If NA or not called in the function, a prompt will appear asking for this data.}

\item{HAUL}{Optional. Default = NA. The haul number, aka the iterative number of this haul (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.}

\item{MODEL_NUMBER}{Optional. Default = NA. The model number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).}

\item{VERSION_NUMBER}{Optional. Default = NA. The version number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).}

\item{SERIAL_NUMBER}{Optional. Default = NA. The serial number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).}

\item{path_in}{Optional. Default = "./., or the local working directory but any path (as a string) may be entered.}

\item{path_out}{Optional. The default is the local working directory but may be specified with a string.}

\item{filename_add}{Optional. Default = "new". This string will be added to the name of the outputed file. Here, you can additional information that may make this file helpful to find later.}

\item{quiet}{Optional logical TRUE/FALSE. Default = FALSE. If FALSE, will print a statement or a pop-up window will let the user know where the file has been saved to.}
}
\value{
.BTH and .BTD files to the path_out directory.
}
\description{
Before running this CTDtoBTD function, you will need to use the CTD laptop to convert the raw CTD data to a .cnv file.
}
\details{
To do this,
\enumerate{
\item use the SBE Data Processing Program, and in the "Run" menu select "1. Data conversion".
\item Under "Program setup file" select the .psa file (should be DataCnv.psa located in CTD folder), under "Instrument configuration file" select the .xmlcon file for the CTD located in the
deployment_xmlcon folder- this is CTD specific, know your CTD's serial number, and under "Input directory" select the .hex file (from the CTD computer) that is specific to the haul that you are missing data.  Click "Start Process".
The conversion program will ask you for outputs. At the prompt, Select and Add "Time, Elapsed, seconds", "Depth, salt water, m", "Temperature, ITS-90, deg C", "Pressure, Strain Gauge, db", and "Conductivity, S/m".
\item This .cnv file must include columns for "Time, Elapsed, seconds", "Depth , salt water, m", "Temperature, ITS-90, deg C", "Pressure, Strain Gauge, db", and "Conductivity, S/m".
\item Press "Start Process" button again and a .cnv file should appear in your selected output directory.  The .cnv file can then be used for this CTDtoBTD() function!
\item Look in your output directory (usually the "Documents" folder, but you can find it using getwd()) for your new .BTD and .BTH files.
}

Note that if there are multiple observations from the CTD per second, that they will be averaged by second (e.g., any observations from seconds 0 to >1.0 will be averaged together).
}
\examples{
CTDtoBTD(
   VESSEL = 94,
   CRUISE = 2001,
   HAUL = 4,
   MODEL_NUMBER = 123, # for example, also can make up
   VERSION_NUMBER = 456, # for example, also can make up
   SERIAL_NUMBER = 789, # for example, also can make up
   path_in = system.file(paste0("exdata/ctd2btd/",
      "SBE19plus_01908103_2021_06_01_94_0004_raw.cnv"), 
       package = "GAPsurvey"),
   path_out = getwd(),
   filename_add = "newctd", 
   quiet = TRUE)
  
 
CTDtoBTD(
   VESSEL = 94,
   CRUISE = 202101,
   HAUL = 107,
   MODEL_NUMBER = "", # for example, also can make up
   VERSION_NUMBER = "", # for example, also can make up
   SERIAL_NUMBER = 8105, # for example, also can make up
   path_in = system.file(paste0("exdata/ctd2btd/",
     "SBE19plus_01908105_2021_06_25_94_0005.cnv"), 
      package = "GAPsurvey"),
   path_out = getwd(),
   filename_add = "newctd", 
   quiet = TRUE)
}
