% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deleteDataEnt}
\alias{deleteDataEnt}
\title{Delete data from DataEnt.mdb}
\usage{
deleteDataEnt(dsnDataEnt, haul, tablename)
}
\arguments{
\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}

\item{haul}{Number of the haul you want to delete}

\item{tablename}{String of the data_ent table you are deleting data from}
}
\value{
Changes to the data_ent.mbd file.
}
\description{
NOTE: Must be run on 32-bit R!
}
\details{
Deletes data from data_ent.mdb for a haul.
}
\examples{
# There will need to be observations in the data_ent.mbd for this example to work. 
dsnTablet <- system.file("exdata/catch/GOA/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/GOA/data_ent.mdb", package = "GAPsurvey")
dMix <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                      tablename = 'MIXTURE')
dMixHead <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                          tablename = 'MIXTURE_HEADER')
dCatch <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                        tablename = 'CATCH')
dCatchHead <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                            tablename = 'CATCH_HEADER')
dLength <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = haul, 
                         tablename = 'LENGTH')
dRawLength <- deleteDataEnt(dsnDataEnt = dsnDataEnt, haul = 
                              haul, tablename = 'RAW_LENGTH')
}
