% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_catch_haul_history}
\alias{get_catch_haul_history}
\title{Find catch data from previous years v2}
\usage{
get_catch_haul_history(
  public_data,
  survey,
  years = NA,
  station,
  grid_buffer = NA
)
}
\arguments{
\item{public_data}{(dataframe) A dataframe containing historical survey data}

\item{years}{(numeric) the years you want returned in the output}

\item{station}{(character) A character string of the current station name (as a grid cell; e.g., "264-85")}

\item{grid_buffer}{(numeric) The number of cells around the current station where you would like to see catches from}
}
\value{
a data.frame of past catches and hauls
}
\description{
Find catch data from previous years v2
}
\examples{
data("public_data", package = "GAPsurvey")

# GOA
get_catch_haul_history(public_data = public_data, 
      survey = "GOA", 
      years = 2019,
      station = "264-150",
      grid_buffer = 3)
get_catch_haul_history(public_data = public_data, 
      survey = "GOA", 
      years = NA,
      station = "264-150",
      grid_buffer = 3)

# AI
get_catch_haul_history(public_data = public_data, 
      survey = "AI", 
      years = 2018,
      station = "324-73",
      grid_buffer = 3)
get_catch_haul_history(public_data = public_data, 
      survey = "AI", 
      years = NA,
      station = "324-73",
      grid_buffer = 3)

# EBS
get_catch_haul_history(public_data = public_data, 
      survey = "EBS", 
      years = 2019:2021, 
      station = "I-13")   
get_catch_haul_history(public_data = public_data, 
      survey = "EBS", 
      station = "I-13")   
}
