% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sunrise_sunset}
\alias{get_sunrise_sunset}
\title{Get sunrise and sunset times by day, latitude, and longitude}
\usage{
get_sunrise_sunset(
  chosen_date,
  latitude = NULL,
  longitude = NULL,
  survey = NULL,
  station = NULL,
  verbose = FALSE,
  timezone = "US/Alaska"
)
}
\arguments{
\item{chosen_date}{Date or charater. Formatted as "YYYY-MM-DD"}

\item{latitude}{Numeric. Fill in only if survey and station are not entered. latitude in either decimal degrees or a character latitude in degrees and decimal minutes}

\item{longitude}{Numeric. Fill in only if survey and station are not entered. Longitude in either decimal degrees or a character longitude in degrees and decimal minutes}

\item{survey}{Character. Fill in only if latitude and longitude are not entered. A character string of the survey you are interested in reivewing. Options are those from public_data$survey, which are "AI", "GOA", "EBS", "NBS", "BSS".}

\item{station}{Character. Fill in only if latitude and longitude are not entered. A character string of the current station name (as a grid cell; e.g., "264-85")}

\item{verbose}{Logical. Default = FALSE. If you would like a readout of what the file looks like in the console, set to TRUE.}

\item{timezone}{Character. Default = "US/Alaska." Other options include: "US/Aleutian"}
}
\value{
Time of sunrise and sunset in text. Also shows a pop-up with sunrise and sunset times.
}
\description{
Get sunrise and sunset times by day, latitude, and longitude
}
\examples{
get_sunrise_sunset(chosen_date = Sys.Date(),
                   latitude = 63.3,
                   longitude = -170.5)
get_sunrise_sunset(chosen_date = as.character(Sys.Date()),
                   latitude = 63.3,
                   longitude = -170.5)
get_sunrise_sunset(chosen_date = Sys.Date(),
                   survey = "EBS",
                   station = "I-13")
get_sunrise_sunset(chosen_date = Sys.Date(),
                   survey = "GOA",
                   station = "7-7")
get_sunrise_sunset(chosen_date = "2023-06-10",
                   survey = "AI",
                   station = "324-73",
                   timezone = "US/Aleutian")
}
