% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sunrise_sunset}
\alias{get_sunrise_sunset}
\title{Get sunrise/sunset}
\usage{
get_sunrise_sunset(chosen_date, latitude, longitude, verbose = FALSE)
}
\arguments{
\item{chosen_date}{Date, formatted as YYYY-MM-DD}

\item{latitude}{Numeric latitude in either decimal degrees or a character latitude in degrees and decimal minutes}

\item{longitude}{Numeric longitude in either decimal degrees or a character longitude in degrees and decimal minutes}

\item{verbose}{Logical. Default = FALSE. If you would like a readout of what the file looks like in the console, set to TRUE.}
}
\value{
Time of sunrise and sunset in text. Also shows a pop-up with sunrise and sunset times.
}
\description{
Get sunrise/sunset
}
\examples{
get_sunrise_sunset(chosen_date = Sys.Date(), latitude = 63, longitude = 170)
}
