% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lengthData}
\alias{lengthData}
\title{Import Length Tablet Files Into DataEnt.mdb}
\usage{
lengthData(haul, dsnTablet, dsnDataEnt)
}
\arguments{
\item{haul}{Number of the haul you want to upload}

\item{dsnTablet}{String for the drive folder (in R notation) where tablet files exist.}

\item{dsnDataEnt}{String for the drive folder (in R notation) - include full file name and extension.}
}
\value{
...
}
\description{
NOTE: Must be run on 32-bit R!
}
\details{
Looks for tablet length files in given DSN and uploads them to LENGTH and RAW_LENGTH in the given data_ent.mdb.
}
