% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{local_racebase}
\alias{local_racebase}
\title{RACEBASE Data Set}
\format{
A data frame with 1613690 rows and 38 variables:
\describe{
\item{\code{cruisejoin}}{integer COLUMN_DESCRIPTION}
\item{\code{hauljoin}}{integer COLUMN_DESCRIPTION}
\item{\code{region}}{character Cruise join var}
\item{\code{vessel}}{integer ID number of the vessel used to collect data for that haul. The column “vessel_id” is associated with the “vessel_name” column. Note that it is possible for a vessel to have a new name but the same vessel id number. For a complete list of vessel ID codes: https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}
\item{\code{cruise}}{integer This is a six-digit number identifying the cruise number of the form: YYYY99 (where YYYY = year of the cruise; 99 = 2-digit number and is sequential; 01 denotes the first cruise that vessel made in this year, 02 is the second, etc.)}
\item{\code{haul}}{integer Haul number}
\item{\code{haul_type}}{integer COLUMN_DESCRIPTION}
\item{\code{performance}}{double COLUMN_DESCRIPTION}
\item{\code{start_time}}{character The date (MM/DD/YYYY) and time (HH:MM) of the beginning of the haul. }
\item{\code{duration}}{double This is the elapsed time between start and end of a haul (decimal hours). }
\item{\code{distance_fished}}{double Distance the net fished (thousandths of kilometers). }
\item{\code{net_width}}{double Measured or estimated distance (meters) between wingtips of the trawl.}
\item{\code{net_measured}}{character COLUMN_DESCRIPTION}
\item{\code{net_height}}{double Measured or estimated distance (meters) between footrope and headrope of the trawl.}
\item{\code{stratum}}{integer RACE database statistical area for analyzing data. Strata were designed using bathymetry and other geographic and habitat-related elements. The strata are unique to each survey series. Stratum of value 0 indicates experimental tows. }
\item{\code{start_latitude}}{double Latitude (one hundred thousandth of a decimal degree) of the start of the haul.}
\item{\code{end_latitude}}{double Latitude (one hundred thousandth of a decimal degree) of the end of the haul.}
\item{\code{start_longitude}}{double Longitude (one hundred thousandth of a decimal degree) of the start of the haul.}
\item{\code{end_longitude}}{double Longitude (one hundred thousandth of a decimal degree) of the end of the haul.}
\item{\code{stationid}}{character Alpha-numeric designation for the station established in the design of a survey. }
\item{\code{gear_depth}}{integer COLUMN_DESCRIPTION}
\item{\code{bottom_depth}}{integer Bottom depth (tenths of a meter).}
\item{\code{bottom_type}}{integer COLUMN_DESCRIPTION}
\item{\code{surface_temperature}}{double COLUMN_DESCRIPTION}
\item{\code{gear_temperature}}{double COLUMN_DESCRIPTION}
\item{\code{wire_length}}{integer COLUMN_DESCRIPTION}
\item{\code{gear}}{integer COLUMN_DESCRIPTION}
\item{\code{accessories}}{integer COLUMN_DESCRIPTION}
\item{\code{subsample}}{integer COLUMN_DESCRIPTION}
\item{\code{abundance_haul}}{character COLUMN_DESCRIPTION}
\item{\code{AreaSwept_km2}}{double COLUMN_DESCRIPTION}
\item{\code{catchjoin}}{integer COLUMN_DESCRIPTION}
\item{\code{species_code}}{integer The species code of the organism associated with the “common_name” and “scientific_name” columns. For a complete species list go to https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}
\item{\code{weight}}{double Weight (thousandths of a kilogram) of individuals in a haul by taxon. }
\item{\code{number_fish}}{double Total number of individuals caught in haul by taxon, represented in whole numbers. }
\item{\code{subsample_code}}{logical COLUMN_DESCRIPTION}
\item{\code{voucher}}{integer COLUMN_DESCRIPTION}
\item{\code{year}}{character Year the survey was conducted in.}
}
}
\usage{
data(local_racebase)
}
\description{
One dataframe containing all the racebase tables (experimental: this might be a bad idea)
}
\details{
DETAILS
}
\examples{
data(local_racebase)
}
\author{
Margaret Siple (margaret.siple AT noaa.gov)
}
\keyword{catch}
\keyword{data}
