% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{local_racebase}
\alias{local_racebase}
\title{RACEBASE Data Set}
\format{
One dataframe containing all the racebase tables (experimental: this might be a bad idea)
\describe{
\item{region }{Cruise join var}
\item{vessel}{Vessel ID, numeric}
\item{cruise}{Six-digit cruise number}
\item{haul}{Haul number}
}
}
\usage{
data(local_racebase)
}
\description{
RACEBASE Data Set
}
\examples{
data(local_racebase)
}
\author{
Margaret Siple (margaret.siple AT noaa.gov)
}
\keyword{catch}
\keyword{data}
