% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{PolySpecies}
\alias{PolySpecies}
\title{PolySpecies Data Set}
\format{
A data frame with 172 obs. of  4 variables:
\describe{
\item{SPECIES_CODE}{Species code}
\item{POLY_SPECIES_CODE}{Poly species code}
\item{SPECIES_NAME}{Species scientific latin name}
\item{COMMON_NAME}{Species common names)}
}
}
\usage{
data(PolySpecies)
}
\description{
PolySpecies Data Set
}
\examples{
data(PolySpecies)
}
\author{
Jason Conner (jason.conner AT noaa.gov)
}
\keyword{catch}
\keyword{data}
