% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{PolySpecies}
\alias{PolySpecies}
\title{PolySpecies Data Set}
\format{
A data frame with 172 rows and 4 variables:
\describe{
\item{\code{SPECIES_CODE}}{integer Species code}
\item{\code{POLY_SPECIES_CODE}}{integer Poly species code}
\item{\code{SPECIES_NAME}}{character Species scientific latin name}
\item{\code{COMMON_NAME}}{character Species common names}
}
}
\usage{
data(PolySpecies)
}
\description{
polynumbers
}
\details{
DETAILS
}
\examples{
data(PolySpecies)
}
\author{
Jason Conner (jason.conner AT noaa.gov)
}
\keyword{catch}
\keyword{data}
