% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{public_data}
\alias{public_data}
\title{Public data from FOSS}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 878805 rows and 33 columns.
}
\usage{
public_data
}
\description{
Public data from FOSS
}
\examples{
data(public_data)
}
\author{
Emily Markowitz (emily.markowitz AT noaa.gov)
}
\keyword{datasets}
