% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{sp_table}
\alias{sp_table}
\title{species data codes}
\format{
An object of class \code{data.frame} with 2754 rows and 2 columns.
}
\usage{
sp_table
}
\description{
species data codes
}
\examples{
data(sp_table)
}
\author{
Margaret Siple (margaret.siple AT noaa.gov)
}
\keyword{datasets}
