% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/species_data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{species data codes}
\format{
A data frame with 172 rows and 4 variables:
\describe{
  \item{\code{species_code}}{integer Species code}
  \item{\code{common_name}}{integer Poly species code}
  \item{\code{scientific_name}}{character Species scientific latin name}
  \item{\code{scientific_name_old}}{character Species scientific latin name used previously}
}

A data frame with 2699 observations on the following 3 variables.
\describe{
  \item{\code{common_name}}{Taxon common name. The common name of the marine organism associated with the 'scientific_name' and 'species_code' columns. For a complete species list, review the [code books](https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual).}
  \item{\code{scientific_name}}{Taxon scientific name. The scientific name of the organism associated with the 'common_name' and 'species_code' columns. For a complete taxon list, review the [code books](https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual).}
  \item{\code{species_code}}{Taxon code. The species code of the organism associated with the 'common_name' and 'scientific_name' columns. For a complete species list, review the [code books](https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual).}#'   }
}
\source{
https://github.com/afsc-gap-products/gap_products
}
\usage{
data(species_data)

data('species_data')
}
\description{
RACEBASE Species Codes and Scientific Names Data Set

species classification table
}
\details{
DETAILS

The Resource Assessment and Conservation Engineering (RACE) Division Groundfish Assessment Program (GAP) of the Alaska Fisheries Science Center (AFSC) conducts fisheries-independent bottom trawl surveys to assess the populations of demersal fish and crab stocks of Alaska.
}
\examples{
data(species_data)
data(species_data)
}
\author{
Sarah Friedman (sarah.friedman AT noaa.gov)
}
\keyword{code}
\keyword{data}
\keyword{scientific}
\keyword{species}
