% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{species data codes}
\format{
A data frame with 172 rows and 4 variables:
\describe{
  \item{\code{species_code}}{integer Species code}
  \item{\code{common_name}}{integer Poly species code}
  \item{\code{scientific_name}}{character Species scientific latin name}
  \item{\code{scientific_name_old}}{character Species scientific latin name used previously}
}
}
\usage{
data(species_data)
}
\description{
RACEBASE Species Codes and Scientific Names Data Set
}
\details{
DETAILS
}
\examples{
data(species_data)
}
\author{
Sarah Friedman (sarah.friedman AT noaa.gov)
}
\keyword{code}
\keyword{data}
\keyword{scientific}
\keyword{species}
