% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_coords.R
\docType{data}
\name{station_coords}
\alias{station_coords}
\title{Station centroid locations for each station from akgfmaps}
\format{
A data frame with 31594 observations on the following 4 variables.
\describe{
  \item{\code{latitude_dd}}{Latitude (decimal degrees). Latitude (one hundred thousandth of a decimal degree).}
  \item{\code{longitude_dd}}{Longitude (decimal degrees). Longitude (one hundred thousandth of a decimal degree).}
  \item{\code{srvy}}{Survey abbreviation. Abbreviated survey names. The column 'srvy' is associated with the 'survey' and 'survey_definition_id' columns. Northern Bering Sea (NBS), Southeastern Bering Sea (EBS), Bering Sea Slope (BSS), Gulf of Alaska (GOA), Aleutian Islands (AI).}
  \item{\code{station}}{Station ID. Alpha-numeric designation for the station established in the design of a survey.}#'   }
}
\source{
https://github.com/afsc-gap-products/akgfmaps
}
\usage{
data('station_coords')
}
\description{
Station centroid coordinates for each station for all surveys, as defined by the {akgfmaps} package.
}
\details{
Find code to create this table in ./inst/run.R
}
\examples{
data(station_coords)
}
\author{
Sean Rohan (sean.rohan AT noaa.gov)
}
\keyword{data}
\keyword{station}
\keyword{survey}
